Fix support for DEB_PYTHON_MODULE_PACKAGE (and drop internal cdbs_python_support_path).
DEB_PYTHON_MODULE_PACKAGE was declared, so any prior declaration was
ignored. Also, it is used in an ifeq construct which gets resolved early
and disqualifies later declarations.  As a result, the option was
impossible to use, and python-distutils.mk supported only packages
matching the default naming rule.

This fix avoids setting the option, and instead conditionally apply its
default to an internal variable.

DEB_PYTHON_MODULE_PACKAGE is still only supported when declared before
including the snippet - and getting rid of the ifeq does not solve this,
as we now use package names in build targets which are also resolved
early.

As a bonus, the internal variable cdbs_python_support_path was
discovered to be unused, so was dropped.


git-svn-id: svn+ssh://svn.debian.org/svn/build-common/trunk@344 19660600-52fe-0310-9875-adc0d7a7b53c