Always warn if deprecated DEB_PYTHON_MODULE_PACKAGE is used.
A deprecation warning is needed unconditionally, as the replacement is
plural, so will break assumpltions on being singular in the future (i.e.
if used to form a build target, see bug#521180).

The classical CDBS deprecation mechanism warns only if a deprecated
variable is redefined.  So replace that with an embedded warning that
triggers whenever the variable is resolved.

git-svn-id: svn+ssh://svn.debian.org/svn/build-common/trunk@360 19660600-52fe-0310-9875-adc0d7a7b53c