Intelligent autotools versions (Closes:#725950)
DEB_AUTO_UPDATE_* can now be a (semicolon-separated) pair
that defines a version range (A<=x<=B).
the version range can be open (e.g. ",12" = '<=12')
or empty (",", unversioned dependency)

if the version is given as a version range, the autotools are invoked
without any version suffix (e.g. "aclocal" instead of "aclocal-1.14"